requirements.txt
streamlit
mysql-connector-python
requests
python-dotenv
generativeai